fixed minor out-of-bounds error in scattering histograms

+++ include/data/Axes.h
@@ -26,6 +26,6 @@ struct Axes {
             return *this;
         }
 
-        int bins;
+        size_t bins;
         double xmin, xmax;
 };
